requirements.txt
colorama==0.4.6
numpy==1.26.3
pillow==10.2.0
